% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMM.R
\name{baseModel}
\alias{baseModel}
\title{Base supervised machine learning models for prediction}
\usage{
baseModel(
  trainData,
  testData,
  classifier = c("randForest", "SVM", "glmnet"),
  predMode = c("classification", "probability", "regression"),
  paramlist
)
}
\arguments{
\item{trainData}{The input training dataset. The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{testData}{The input test dataset. The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{classifier}{Machine learning classifiers. 
Available options are c('randForest', 'SVM', 'glmnet').}

\item{predMode}{The prediction mode. Available options are
c('classification', 'probability', 'regression'). 
'probability' is currently only for 'randForest'.}

\item{paramlist}{A set of model parameters defined in an R list object. 
See more details for each individual model.}
}
\value{
Based on a given machine learning, the predicted score/output will be 
estimated for the test data.
}
\description{
Prediction using different supervised machine learning models.
}
\examples{
 
## Load data  
methylfile <- system.file('extdata', 'methylData.rds', package='BioMM')  
methylData <- readRDS(methylfile)  
dataY <- methylData[,1]
## select a subset of genome-wide methylation data at random
methylSub <- data.frame(label=dataY, methylData[,c(2:2001)])  
trainIndex <- sample(nrow(methylSub), 16)
trainData = methylSub[trainIndex,]
testData = methylSub[-trainIndex,]
library(ranger) 
set.seed(123)
predY <- baseModel(trainData, testData, 
                   classifier='randForest',  
                   predMode='classification', 
                   paramlist=list(ntree=300, nthreads=20)) 
print(table(predY)) 
testY <- testData[,1]
accuracy <- classifiACC(dataY=testY, predY=predY)
print(accuracy)  
}
\author{
Junfang Chen
}
