% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMM.R
\name{BioMMstage2pred}
\alias{BioMMstage2pred}
\title{Prediction performance for stage-2 data using supervised machine learning}
\usage{
BioMMstage2pred(trainData, testData, resample = "CV", dataMode,
  repeatA = 1, repeatB = 1, nfolds, FSmethod, cutP, fdr,
  FScore = MulticoreParam(), classifier, predMode, paramlist,
  innerCore = MulticoreParam())
}
\arguments{
\item{trainData}{The input training dataset (stage-2 data). The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{testData}{The input test dataset (stage-2 data). The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{resample}{The resampling methods. Valid options are 'CV' and 'BS'. 
'CV' for cross validation and 'BS' for bootstrapping resampling.
The default is 'CV'.}

\item{dataMode}{The mode of data used. 'subTrain' or 'allTrain'.}

\item{repeatA}{The number of repeats N is used during resampling prediction. 
The default is 1.}

\item{repeatB}{The number of repeats N is used for test data prediction. 
The default is 1.}

\item{nfolds}{The number of folds is defined for cross validation.}

\item{FSmethod}{Feature selection methods. Available options are 
c(NULL, 'positive', 'wilcox.test', 'cor.test', 'chisq.test', 'posWilcox', 
or 'top10pCor').}

\item{cutP}{The cutoff used for p value thresholding.
Commonly used cutoffs are c(0.5, 0.1, 0.05, 0.01, etc). The default is 0.05.}

\item{fdr}{Multiple testing correction method. Available options are 
c(NULL, 'fdr', 'BH', 'holm', etc). 
See also \code{\link[stats]{p.adjust}}. The default is NULL.}

\item{FScore}{The number of cores used for feature selection if parallel 
computing needed.}

\item{classifier}{Machine learning classifiers.}

\item{predMode}{The prediction mode. Available options are 
c('probability', 'classification', 'regression').}

\item{paramlist}{A set of model parameters defined in an R list object.}

\item{innerCore}{The number of cores used for computation.}
}
\value{
The CV or BS prediction performance for stage-2 training data and 
test set prediction performance for stage-2 test data if the test set is 
given.
}
\description{
Prediction performance for reconstructed stage-2 data using supervised 
machine learning with feature selection methods.
}
\details{
Stage-2 prediction is performed typically using positively  
correlated features. Since negative associations likely reflect random 
effects in the underlying data
}
\references{
Perlich, C., & Swirszcz, G. (2011). On cross-validation and  
stacking: Building seemingly predictive models on random data. ACM SIGKDD    
Explorations Newsletter, 12(2), 11-15.
}
\author{
Junfang Chen
}
