% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{cirPlot4pathway}
\alias{cirPlot4pathway}
\title{Circular plot for a set of pathways}
\usage{
cirPlot4pathway(datalist, topPathID, core, fileName = NULL)
}
\arguments{
\item{datalist}{The input data list containing ordered 
collections of matrices.}

\item{topPathID}{A predefined pathway IDs.}

\item{core}{The number of cores used for computation. (Default: 10)}

\item{fileName}{The plot file name. (Default: 'cpgTopGOplot.png')}
}
\value{
An output image file.
}
\description{
Plot the individual CpGs within a given set of pathways.
The significance of the CpGs are illustrated by the negative log P value.
}
\details{
Top 10 or 20 pathways are usually suggested to be visualized.
}
\seealso{
\code{\link{omics2pathlist}}.
}
