% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{getMetrics}
\alias{getMetrics}
\title{Compute the evaluation metrics}
\usage{
getMetrics(dataY, predY)
}
\arguments{
\item{dataY}{The observed outcome.}

\item{predY}{The predicted outcome.}
}
\value{
A set of metrics for model evaluation: pv, cor, AUC, ACC and R2.
}
\description{
Compute the evaluation metrics in the classification setting: 
P value based on chi-square test (pv), pearson correlation coefficient 
(cor), area under curve (AUC), classification accuracy (ACC) and 
the pseudo R square (R2).
}
\details{
If all samples are predicted into one class, then we assign 
R2=0, cor=0, and AUC=0.5.
}
\examples{
 
## Load data  
methylfile <- system.file("extdata", "methylData.rds", package="BioMM")  
methylData <- readRDS(methylfile)   
dataY <- methylData[,1]
methylSub <- data.frame(label=dataY, methylData[,c(2:2001)])  
library(ranger) 
library(rms)
predY <- predByCV(methylSub, repeats=1, nfolds=10,   
                  FSmethod=NULL, cutP=0.1, 
                  fdr=NULL, FScore=1, 
                  classifier="randForest",
                  predMode="classification", 
                  paramlist=list(ntree=300, nthreads=1),
                  innerCore=1)   
accuracy <- getMetrics(dataY=dataY, predY=predY)
print(accuracy)  
}
\author{
Junfang Chen
}
