% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMM.R
\name{predByCV}
\alias{predByCV}
\title{Cross validation prediction by supervised machine learning and feature 
selection}
\usage{
predByCV(
  data,
  repeats,
  nfolds,
  FSmethod,
  cutP,
  fdr,
  FScore = MulticoreParam(),
  classifier,
  predMode,
  paramlist,
  innerCore = MulticoreParam()
)
}
\arguments{
\item{data}{The input dataset. The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{repeats}{The number of repeats used for cross validation.
Repeated cross validation is performed if N >=2.}

\item{nfolds}{The number of folds is defined for cross validation.}

\item{FSmethod}{Feature selection methods. Available options are 
c(NULL, 'positive', 'wilcox.test', 'cor.test', 'chisq.test', 'posWilcox', 
or 'top10pCor').}

\item{cutP}{The cutoff used for p value thresholding.
Commonly used cutoffs are c(0.5, 0.1, 0.05, 0.01, etc). The default is 0.05.}

\item{fdr}{Multiple testing correction method. Available options are 
c(NULL, 'fdr', 'BH', 'holm', etc). 
See also \code{\link[stats]{p.adjust}}. The default is NULL.}

\item{FScore}{The number of cores used for feature selection if parallel 
computing needed.}

\item{classifier}{Machine learning classifiers.}

\item{predMode}{The prediction mode. Available options are 
c('probability', 'classification', 'regression').}

\item{paramlist}{A set of model parameters defined in an R list object.}

\item{innerCore}{The number of cores used for computation.}
}
\value{
The predicted cross validation output.
}
\description{
Prediction by supervised machine learning models using cross validation 
along with feature selection methods.
}
\examples{
 
## Load data  
methylfile <- system.file('extdata', 'methylData.rds', package='BioMM')  
methylData <- readRDS(methylfile)   
dataY <- methylData[,1]
## select a subset of genome-wide methylation data at random
methylSub <- data.frame(label=dataY, methylData[,c(2:2001)])  
library(ranger) 
library(BiocParallel)
param1 <- MulticoreParam(workers = 1)
param2 <- MulticoreParam(workers = 20)
predY <- predByCV(methylSub, repeats=1, nfolds=10,   
                  FSmethod=NULL, cutP=0.1, 
                  fdr=NULL, FScore=param1, 
                  classifier='randForest',
                  predMode='classification', 
                  paramlist=list(ntree=300, nthreads=1),
                  innerCore=param2)
dataY <- methylData[,1]
accuracy <- classifiACC(dataY=dataY, predY=predY)
print(accuracy)  
}
