% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMM.R
\name{BioMM}
\alias{BioMM}
\title{BioMM end-to-end prediction}
\usage{
BioMM(trainData, testData, pathlistDB, featureAnno, restrictUp,
  restrictDown, minPathSize, supervisedStage1 = TRUE, typePCA,
  resample1 = "BS", resample2 = "CV", repeatA1 = 100, repeatA2 = 1,
  repeatB1 = 20, repeatB2 = 1, nfolds = 10, FSmethod1, FSmethod2,
  cutP1, cutP2, fdr2, FScore = MulticoreParam(), classifier, predMode,
  paramlist, innerCore = MulticoreParam())
}
\arguments{
\item{trainData}{The input training dataset. The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{testData}{The input test dataset. The first column
is the label or the output. For binary classes, 
0 and 1 are used to indicate the class member.}

\item{pathlistDB}{A list of pathways with pathway IDs and their 
corresponding genes ('entrezID' is used). This is only used for 
pathway-based stratification (only \code{stratify} is 'pathway').}

\item{featureAnno}{The annotation data stored in a data.frame for probe 
mapping. It must have at least two columns named 'ID' and 'entrezID'.  
If it's NULL, then the input probe is from the transcriptomic data. 
(Default: NULL)}

\item{restrictUp}{The upper-bound of the number of probes or genes in each
biological stratified block.}

\item{restrictDown}{The lower-bound of the number of probes or genes in 
each biological stratified block.}

\item{minPathSize}{The minimal defined pathway size after mapping your 
own data to GO database. This is only used for 
pathway-based stratification (only \code{stratify} is 'pathway').}

\item{supervisedStage1}{A logical value. If TRUE, then supervised learning 
models are applied; if FALSE, unsupervised learning.}

\item{typePCA}{the type of PCA. Available options are c('regular', 
'sparse').}

\item{resample1}{The resampling methods at stage-1. Valid options are 
'CV' and 'BS'. 'CV' for cross validation and 'BS' for bootstrapping 
resampling. The default is 'BS'.}

\item{resample2}{The resampling methods at stage-2. Valid options are 'CV' 
and 'BS'. 'CV' for cross validation and 'BS' for bootstrapping resampling.
The default is 'CV'.}

\item{repeatA1}{The number of repeats N is used during resampling procedure.
Repeated cross validation or multiple boostrapping is performed if N >=2. 
One can choose 10 repeats for 'CV' and 100 repeats for 'BS'.}

\item{repeatA2}{The number of repeats N is used during resampling 
prediction. The default is 1 for 'CV'.}

\item{repeatB1}{The number of repeats N is used for generating stage-2 test 
data prediction scores. The default is 20.}

\item{repeatB2}{The number of repeats N is used for test data prediction. 
The default is 1.}

\item{nfolds}{The number of folds is defined for cross validation. 
The default is 10.}

\item{FSmethod1}{Feature selection methods at stage-1. Available options 
are c(NULL, 'positive', 'wilcox.test', 'cor.test', 'chisq.test', 
'posWilcox').}

\item{FSmethod2}{Feature selection methods at stage-2. Available options
are c(NULL, 'positive', 'wilcox.test', 'cor.test', 'chisq.test', 
'posWilcox').}

\item{cutP1}{The cutoff used for p value thresholding at stage-1.
Commonly used cutoffs are c(0.5, 0.1, 0.05, 0.01, etc).}

\item{cutP2}{The cutoff used for p value thresholding at stage-2.}

\item{fdr2}{Multiple testing correction method at stage-2. 
Available options are c(NULL, 'fdr', 'BH', 'holm', etc). 
See also \code{\link[stats]{p.adjust}}. The default is NULL.}

\item{FScore}{The number of cores used for feature selection.}

\item{classifier}{Machine learning classifiers at both stages.}

\item{predMode}{The prediction mode at both stages. Available options are 
c('probability', 'classification', 'regression').}

\item{paramlist}{A list of model parameters at both stages.}

\item{innerCore}{The number of cores used for computation.}
}
\value{
The CV or BS prediction performance for the training data and 
test set prediction performance if \code{testData} is given.
}
\description{
End-to-end prediction by BioMM framework using either supervised or 
unsupervised learning at stage-1, then supervised learning at stage-2.
}
\details{
Stage-2 training data can be learned either using bootstrapping 
or cross validation resampling methods in the supervised learning settting.
Stage-2 test data is learned via independent test set prediction.
}
\examples{
 
## Load data    
methylfile <- system.file('extdata', 'methylData.rds', package='BioMM')  
methylData <- readRDS(methylfile)    
## Annotation files for Mapping CpGs into chromosome  
probeAnnoFile <- system.file('extdata', 'cpgAnno.rds', package='BioMM')  
probeAnno <- readRDS(file=probeAnnoFile)   
supervisedStage1=TRUE
classifier <- 'randForest'
predMode <- 'classification'
paramlist <- list(ntree=300, nthreads=30)   
library(BiocParallel)
library(ranger)
param1 <- MulticoreParam(workers = 2)
param2 <- MulticoreParam(workers = 20)
## Not Run 
## result <- BioMM(trainData=methylData, testData=NULL,
##                 pathlistDB, featureAnno=probeAnno, 
##                 restrictUp=10, restrictDown=200, minPathSize=10, 
##                 supervisedStage1, typePCA='regular', 
##                 resample1='BS', resample2='CV', 
##                 repeatA1=20, repeatA2=1, repeatB1=20, repeatB2=1, 
##                 nfolds=10, FSmethod1=NULL, FSmethod2=NULL, 
##                 cutP1=0.1, cutP2=0.1, fdr2=NULL, FScore=param1, 
##                 classifier, predMode, paramlist, innerCore=param2)
}
\references{
Chen, J., & Schwarz, E. (2017). BioMM: Biologically-informed 
Multi-stage Machine learning for identification of epigenetic fingerprints. 
arXiv preprint arXiv:1712.00336.

Perlich, C., & Swirszcz, G. (2011). On cross-validation and  
stacking: Building seemingly predictive models on random data. ACM SIGKDD    
Explorations Newsletter, 12(2), 11-15.
}
\seealso{
\code{\link{BioMMreconData}}; \code{\link{BioMMstage1pca}}; 
\code{\link{BioMMstage2pred}}
}
