% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{classifiACC}
\alias{classifiACC}
\title{Compute the classification accuracy}
\usage{
classifiACC(dataY, predY)
}
\arguments{
\item{dataY}{The observed outcome.}

\item{predY}{The predicted outcome.}
}
\value{
The classification accuracy in terms of percentage.
}
\description{
Compute the classification accuracy for the binary classification problem.
}
\examples{
 
## Load data  
methylfile <- system.file('extdata', 'methylData.rds', package='BioMM')  
methylData <- readRDS(methylfile) 
dataY <- methylData[,1]
methylSub <- data.frame(label=dataY, methylData[,c(2:1001)])  
library(ranger)  
library(parallel)
param1 <- MulticoreParam(workers = 1) 
param2 <- MulticoreParam(workers = 10) 
predY <- predByCV(methylSub, repeats=1, nfolds=10,   
                  FSmethod=NULL, cutP=0.1, 
                  fdr=NULL, FScore=param1, 
                  classifier='randForest',
                  predMode='classification', 
                  paramlist=list(ntree=300, nthreads=1),
                  innerCore=param2)  
accuracy <- classifiACC(dataY=dataY, predY=predY)
print(accuracy)  
}
\author{
Junfang Chen
}
