% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{plotRankedFeature}
\alias{plotRankedFeature}
\title{Plot top outcome-associated features}
\usage{
plotRankedFeature(data, posF = TRUE, topF = 10, blocklist,
  rankMetric = c("AUC", "ACC", "R2", "size"), colorMetric = c("AUC",
  "ACC", "R2", "size"), core = MulticoreParam(), fileName = NULL)
}
\arguments{
\item{data}{The input stage-2 data (either data.frame or matrix). 
Rows are the samples, columns are pathway names,  
except that the first column is the label (the outcome).}

\item{posF}{A logical value indicating if only positively outcome-associated
features should be used. (Default: TRUE)}

\item{topF}{The top ranked number of features at stage-2 (\code{topF} >= 2).
(Default: 10)}

\item{blocklist}{A list of matrices with block IDs as the associated list 
member names. The block IDs identical to the stage-2 feature names.  
For each matrix, rows are the samples and columns are the probe names,  
except that the first column is named 'label'. See also 
\code{\link{omics2pathlist}}.}

\item{rankMetric}{A string representing the metrics used for ranking. 
Valid options are c('AUC', 'ACC', 'R2', 'size').
'size' is the block size.}

\item{colorMetric}{A string representing the metric used to color the plot. 
Valid options are c('AUC', 'ACC', 'R2', 'size').
'size' is the block size.}

\item{core}{The number of cores used for computation. (Default: 10)}

\item{fileName}{The plot file name. (Default: 'plottopF.png')}
}
\value{
An output image file.
}
\description{
Plot top ranked outcome-associated features from stage-2 data. 
The ranking criteria are based on metrics such as Nagelkerke pseudo 
R-square.
}
\details{
If the argument \code{posF} is TRUE, 
and no positively outcome-associated features are present in stage-2 data 
, then an error is reported. In addition, if \code{topF} is bigger than
the number of positively outcome-associated features, an error is returned.
}
\references{
Perlich, C., & Swirszcz, G. (2011). On cross-validation and  
stacking: Building seemingly predictive models on random data. ACM SIGKDD    
Explorations Newsletter, 12(2), 11-15.
}
\seealso{
\code{\link{omics2pathlist}}.
}
