% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{plotRankedFeature}
\alias{plotRankedFeature}
\title{Plot top outcome-associated features}
\usage{
plotRankedFeature(
  data,
  posF = TRUE,
  topF = 10,
  blocklist,
  binarize = FALSE,
  rankMetric = c("AUC", "R2", "Zscore", "negPlogit", "negPwilcox", "size"),
  colorMetric = c("AUC", "R2", "Zscore", "negPlogit", "negPwilcox", "size"),
  core = MulticoreParam(),
  pathTitle = "GO pathways",
  fileName = NULL
)
}
\arguments{
\item{data}{The input stage-2 data (either data.frame or matrix). 
Rows are the samples, columns are pathway names,  
except that the first column is the label (the outcome).}

\item{posF}{A logical value indicating if only positively outcome-associated
features should be used. (Default: TRUE)}

\item{topF}{The top ranked number of features at stage-2 (\code{topF} >= 2).
(Default: 10)}

\item{blocklist}{A list of matrices with block IDs as the associated list 
member names. The block IDs identical to the stage-2 feature names.  
For each matrix, rows are the samples and columns are the probe names,  
except that the first column is named 'label'. See also 
\code{\link{omics2pathlist}}.}

\item{binarize}{A logical value indicating if the individual features under
investigation should be binarized. The default is FALSE, which provides the 
estimated class probabilities for each pathway-level feature. If TRUE, then
the binary output is given for each feature.}

\item{rankMetric}{A string representing the metrics used for ranking. 
Valid options are c("AUC", "R2", "Zscore", "negPlogit", "negPwilcox").
"negPlogit" denotes the negative log P value from the logistic regression 
and "negPwilcox" means the negative log P value based on the Wilcoxon test. 
"size" is the block size.}

\item{colorMetric}{A string representing the metric used to color the plot. 
Valid options are c("AUC", "R2", "Zscore", "negPlogit", "negPwilcox").
"negPlogit" denotes the negative log P value from the logistic regression 
and "negPwilcox" means the negative log P value based on wilcoxon test. 
"size" is the block size.}

\item{core}{The number of cores used for computation. (Default: 10)}

\item{pathTitle}{A string indicating the name of pathway under investigation.}

\item{fileName}{The plot file name. (Default: 'plottopF.png')}
}
\value{
An output image file and the summary statistics of the top pathways.
}
\description{
Plot top ranked outcome-associated features from stage-2 data. 
The ranking criteria are based on metrics such as Nagelkerke pseudo 
R-square.
}
\details{
If the argument \code{posF} is TRUE, 
and no positively outcome-associated features are present in stage-2 data 
, then an error is reported. In addition, if \code{topF} is bigger than
the number of positively outcome-associated features, an error is returned.
}
\references{
Perlich, C., & Swirszcz, G. (2011). On cross-validation and  
stacking: Building seemingly predictive models on random data. ACM SIGKDD    
Explorations Newsletter, 12(2), 11-15.
}
\seealso{
\code{\link{omics2pathlist}}.
}
