% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultReport.R
\name{plotVarExplained}
\alias{plotVarExplained}
\title{Plot data summary statistics}
\usage{
plotVarExplained(data, posF = TRUE, stratify = c("pathway"),
  core = MulticoreParam(), fileName = NULL)
}
\arguments{
\item{data}{The input dataset (either data.frame or matrix). 
Rows are the samples, columns are the probes/genes, except that 
the first column is the label (the outcome).}

\item{posF}{A logical value indicating if only positively outcome-associated
features should be used. (Default: TRUE)}

\item{stratify}{A string. Pathway based stratification method to generate 
\code{blocklist}.}

\item{core}{The number of cores used for computation. (Default: 1)}

\item{fileName}{The file name specified for the plot. If it is not NULL,
then the plot will be generated. The plot will project the data on the 
first two components. (Default: 'R2explained.png')}
}
\value{
An output image file with '.png' format.
}
\description{
Plot data summary statistics in terms of the proportion of 
variance explained.
}
\references{
Yu, Guangchuang, et al. 'clusterProfiler: an R package for 
comparing biological themes among gene clusters.' Omics: a journal of 
integrative biology 16.5 (2012): 284-287.

Perlich, C., & Swirszcz, G. (2011). On cross-validation and  
stacking: Building seemingly predictive models on random data. ACM SIGKDD    
Explorations Newsletter, 12(2), 11-15.
}
